+++ R/comparison-summary.R
@@ -6,7 +6,6 @@ summary.rcmdcheck_comparison <- function(object, ...) {
 }
 
 #' @export
-#' @importFrom crayon bgRed white green
 
 print.rcmdcheck_comparison_summary <- function(x, ...) {
   object <- x[[1]]
@@ -27,7 +26,7 @@ print.rcmdcheck_comparison_summary <- function(x, ...) {
     change_summary(object$cmp, "error"), " | ",
     change_summary(object$cmp, "warning"), " | ",
     change_summary(object$cmp, "note"),
-    style = make_style("darkgrey")
+    style = darkgrey
   )
 
   invisible(x)